Move server/auth.go into server/api_server.go. Also rename auth and enterprise packages that don't agree with their directory
