+++ src/server/enterprise/server/api_server.go
@@ -1,4 +1,4 @@
-package enterprise
+package server
 
 import (
 	"crypto"
