+++ src/server/enterprise/server/enterprise_test.go
@@ -1,4 +1,4 @@
-package enterprise
+package server
 
 import (
 	"testing"
